Move fluent search_registered_models
